+++ utils/ConfigurationManager.ts
