+++ utils/DeviceData.ts
